define __hva2spa__ and __spa2hva__ in terms of hva2spa and spa2hva to cut redundant logic. lots of casting fix-ups to support this change
git-svn-id: https://sparrow.ece.cmu.edu/svn/hyp@3062 1c601a1c-44f4-4d93-90ac-4b632a93592e

+++ emhf/trunk/code/x86/runtime/apic_svm.c
@@ -177,8 +177,8 @@ u32 svm_lapic_access_handler(VCPU *vcpu, u32 paddr, u32 errorcode){
       //change LAPIC physical address NPT mapping to point to physical 
       //address of virtual_LAPIC_base
       //printf("\nvirtual_LAPIC_base, v=0x%08x, p=0x%08x",  
-      //  (u32)virtual_LAPIC_base, __hva2spa__((u32)virtual_LAPIC_base));
-      npt_changemapping(vcpu, g_svm_lapic_base, __hva2spa__((u32)g_svm_virtual_LAPIC_base), (u64)(_PAGE_PRESENT | _PAGE_RW | _PAGE_USER));
+      //  (u32)virtual_LAPIC_base, __hva2spa__(virtual_LAPIC_base));
+      npt_changemapping(vcpu, g_svm_lapic_base, __hva2spa__(g_svm_virtual_LAPIC_base), (u64)(_PAGE_PRESENT | _PAGE_RW | _PAGE_USER));
       vmcb->tlb_control = TLB_CONTROL_FLUSHALL;  
 
     }else{
@@ -205,8 +205,8 @@ u32 svm_lapic_access_handler(VCPU *vcpu, u32 paddr, u32 errorcode){
       //change LAPIC physical address NPT mapping to point to physical 
       //address of virtual_LAPIC_base
       //printf("\nvirtual_LAPIC_base, v=0x%08x, p=0x%08x",  
-      //  (u32)virtual_LAPIC_base, __hva2spa__((u32)virtual_LAPIC_base));
-      npt_changemapping(vcpu, g_svm_lapic_base, __hva2spa__((u32)g_svm_virtual_LAPIC_base), (u64)(_PAGE_PRESENT | _PAGE_RW | _PAGE_USER));
+      //  (u32)virtual_LAPIC_base, __hva2spa__(virtual_LAPIC_base));
+      npt_changemapping(vcpu, g_svm_lapic_base, __hva2spa__(g_svm_virtual_LAPIC_base), (u64)(_PAGE_PRESENT | _PAGE_RW | _PAGE_USER));
       vmcb->tlb_control = TLB_CONTROL_FLUSHALL;  
 
     }else{
